server/requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
clickclick==20.10.2
connexion==2.7.0
Flask==1.1.2
idna==2.10
inflection==0.5.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==2.6.0
ldap3==2.8.1
MarkupSafe==1.1.1
openapi-spec-validator==0.2.9
psycopg2-binary==2.8.6
pyasn1==0.4.8
python-dateutil==2.6.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.20
swagger-ui-bundle==0.0.8
urllib3==1.25.11
uWSGI==2.0.19.1
Werkzeug==1.0.1
